Haste.GAPI

Documentation

data Config

Google API config

Constructors

Config

Fields

clientID :: String

apiKey :: String

scopes :: String

immediate :: Bool

data OAuth2Token

OAuth2 Token

Constructors

OA2Success

Fields

accessToken :: String

expiresIn :: String

state :: String

OA2Error

Fields

errorMsg :: String

state :: String

data Library

Represents a GAPILibrary with a name and a version

Constructors

Lib String String

data Promise

type Reason = JSAny

type Response = JSAny

withGAPI :: Config -> (OAuth2Token -> IO ()) -> IO ()

oa2success :: OAuth2Token -> Bool

Returns true if the token represents a successful authentication

getToken :: IO OAuth2Token

Returns the token from the current GAPI state

loadLibrary :: Library -> Promise -> IO ()

Loads a library using a promise

withLibrary :: Library -> IO () -> IO ()

Loads a library, and then executes the second argument as a callback

gapiError :: (JSString -> IO ()) -> Reason -> IO ()

Default error handler in promises

fromResult :: FromResult result => Result -> RequestM result