Haste.GAPI.GPlus

Documentation

data Person

Constructors

Person

Fields

kind :: String

etag :: ETag

nickname :: String

occupation :: String

skills :: String

birthday :: String

gender :: String

emails :: [(EMail, EMailType)]

urls :: [Site]

objectType :: String

id :: UserID

displayName :: String

name :: Name

tagline :: String

braggingRights :: String

aboutMe :: String

relationshipStatus :: String

url :: String

image :: Image

organisations :: [Organization]

placesLived :: [Place]

isPlusUser :: Bool

language :: String

ageRange :: AgeRange

plusOneCount :: Int

circledByCount :: Int

verified :: Bool

cover :: Cover

domain :: String

data Name

Name type of a Google+ user

Constructors

Name

Fields

formatted :: String

familyName :: String

givenName :: String

middleName :: String

honorificPrefix :: String

honorificSuffix :: String

data Site

An URL of a Google+ user,

Constructors

Site

Fields

label :: String

siteHref :: URL

siteType :: URLType

type EMail = String

An e-mail address

type EMailType = String

Type of e-mail address.

type URL = String

An URL

type URLType = String

Type of URL, like contributor (user is contributing to this site) or website.

data Image

An image wrapped URL

Constructors

Image URL

data Organization

Constructors

Organization

Fields

orgName :: String

orgDepartment :: String

orgTitle :: String

orgType :: String

orgStartDate :: String

orgEndDate :: String

orgLocation :: String

orgDescription :: String

orgPrimary :: Bool

data AgeRange

Constructors

AgeRange

Fields

min :: Int

max :: Int

data Cover

Constructors

Cover

Fields

layout :: String

coverPhoto :: CoverPhoto

coverInfo :: CoverInfo

peopleGet :: UserID -> RequestM Person

Fetches a user by ID

peopleSearch :: String -> Params -> RequestM [Person]

Search after users given a query string

Optional parameters:

language

language to search with. See Google API documentation for valid language codes

maxResult

Maximum number of people to include in the response. Acceptable values of maxResult ranges between 1-50, default is 25.

pageToken

Token used for pagination in large result sets.

peopleListByActivity :: ActivityID -> Collection -> Params -> RequestM [Person]

List users by activity

Valid collections are "plusoners" and "resharers".

Optional parameters:

maxResult

Maximum number of people to include in the response. Acceptable values of maxResult ranges between 1-100, default is 20.

pageToken

Token used for pagination in large result sets.

peopleList :: Collection -> Params -> RequestM [Person]

List people in a specific collection

Accepted collections are:

connected

The list of visible people in the authenticated user's circles who also use the requesting app. This list is limited to users who made their app activities visible to the authenticated user.

visible

The list of people who this user has added to one or more circles, limited to the circles visible to the requesting application.

Optional parameters:

maxResult

Maximum number of people to include in the response. Acceptable values of maxResult ranges between 1-100, default is 100.

orderBy

The order to return people in. Valid inputs are "alphabetical" and "best"

pageToken

Token used for pagination in large result sets.