Haste.GAPI.Request

Documentation

data Promise

newtype RequestM a

Type responsible for executing sensible requests. Example: This code will execute a request with a path and a set of parameters (params), uses fields for fetching the fields "name" and "email" from the first response and creating params' with these fields, and uses them as argument to the second call to request. do response <- request aPath params params' <- fields ["name", "email"] response response' <- request aPath' params'

Constructors

Req

Fields

unR :: CIO (Either Error a)

type Reason = JSAny

type Response = JSAny

data Params

Parameters for a GAPI request

Constructors

Params [(String, String)]

data Request

Request with parameters and everything

Constructors

Request

Fields

path :: String

method :: String

params :: Params

headers :: String

body :: String

newtype Result

Constructors

Result

Fields

v :: JSAny

type Path = String

Path for requests

class FromResult result where

Methods

fromResult :: Result -> RequestM result

gapiError :: (JSString -> IO ()) -> Reason -> IO ()

Default error handler in promises

rawRequest :: String -> Params -> Request

Creates a request by manually entering request path and parameters

withRequest :: Request -> Promise -> IO ()

Applies request and executes the given Promise.

cRequest :: Request -> CIO JSAny

Executes a request, blocking while waiting for the result and then return the finished equation

req :: RequestM () -> IO ()

Executes a request

request :: Path -> Params -> RequestM Result

Creates a request

debugResult :: Result -> RequestM ()

has :: Result -> String -> RequestM Bool

Checks whether your result has a certain field

get :: FromAny a => Result -> String -> RequestM a

Gets a field from a result, on missing field RequestM will fail.

parp :: Result -> [String] -> RequestM Params

Maps a couple of keys out to a new set of params

liftIO :: MonadIO m => forall a. IO a -> m a

deepGet :: FromAny a => Result -> String -> RequestM (Maybe a)

like get, but will look at deeper objects given a . separated string.

defFromResult :: FromAny a => Result -> RequestM a